+++ src/main/java/com/cheocharm/MapZ/usergroup/domain/InvitationStatus.java
@@ -1,4 +1,4 @@
-package com.cheocharm.MapZ.usergroup;
+package com.cheocharm.MapZ.usergroup.domain;
 
 import lombok.Getter;
